Move types from namespace Steeltoe.Management.Diagnostics in Management.Abstractions project into Management.Endpoint project
