+++ src/Management/src/Endpoint/Actuators/Metrics/Diagnostics/IDiagnosticsManager.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the Apache 2.0 License.
 // See the LICENSE file in the project root for more information.
 
-namespace Steeltoe.Management.Diagnostics;
+namespace Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics;
 
 public interface IDiagnosticsManager
 {
