+++ src/Management/src/Endpoint/PublicAPI.Unshipped.txt
@@ -1,5 +1,6 @@
 #nullable enable
 abstract Steeltoe.Management.Endpoint.Middleware.EndpointMiddleware<TArgument, TResult>.InvokeEndpointHandlerAsync(Microsoft.AspNetCore.Http.HttpContext! context, System.Threading.CancellationToken cancellationToken) -> System.Threading.Tasks.Task<TResult>!
+abstract Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.DiagnosticObserver.ProcessEvent(string! eventName, object? value) -> void
 const Steeltoe.Management.Endpoint.Actuators.Health.Availability.ApplicationAvailability.LivenessKey = "Liveness" -> string!
 const Steeltoe.Management.Endpoint.Actuators.Health.Availability.ApplicationAvailability.ReadinessKey = "Readiness" -> string!
 override Steeltoe.Management.Endpoint.Actuators.Health.Availability.AvailabilityState.ToString() -> string!
@@ -292,6 +293,20 @@ Steeltoe.Management.Endpoint.Actuators.Loggers.LoggersResponse.HasError.get -> b
 Steeltoe.Management.Endpoint.Actuators.Loggers.LoggersResponse.Levels.get -> System.Collections.Generic.IList<string!>!
 Steeltoe.Management.Endpoint.Actuators.Loggers.LoggersResponse.Loggers.get -> System.Collections.Generic.IDictionary<string!, Steeltoe.Management.Endpoint.Actuators.Loggers.LoggerLevels!>!
 Steeltoe.Management.Endpoint.Actuators.Loggers.LoggersResponse.LoggersResponse(System.Collections.Generic.IList<string!>! levels, System.Collections.Generic.IDictionary<string!, Steeltoe.Management.Endpoint.Actuators.Loggers.LoggerLevels!>! loggers, System.Collections.Generic.IDictionary<string!, Steeltoe.Management.Endpoint.Actuators.Loggers.LoggerGroup!>! groups) -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.DiagnosticObserver
+Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.DiagnosticObserver.DiagnosticObserver(string! name, string! listenerName, Microsoft.Extensions.Logging.ILoggerFactory! loggerFactory) -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.DiagnosticObserver.Dispose() -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.DiagnosticObserver.ListenerName.get -> string!
+Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.DiagnosticObserver.ObserverName.get -> string!
+Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.DiagnosticObserver.Subscribe(System.Diagnostics.DiagnosticListener! listener) -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.IDiagnosticObserver
+Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.IDiagnosticObserver.ObserverName.get -> string!
+Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.IDiagnosticObserver.Subscribe(System.Diagnostics.DiagnosticListener! listener) -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.IDiagnosticsManager
+Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.IDiagnosticsManager.Start() -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.IDiagnosticsManager.Stop() -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.IRuntimeDiagnosticSource
+Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.IRuntimeDiagnosticSource.AddInstrumentation() -> void
 Steeltoe.Management.Endpoint.Actuators.Metrics.EndpointServiceCollectionExtensions
 Steeltoe.Management.Endpoint.Actuators.Metrics.IMetricsEndpointHandler
 Steeltoe.Management.Endpoint.Actuators.Metrics.MetricSample
@@ -308,6 +323,28 @@ Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsEndpointOptions.MaxHistogr
 Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsEndpointOptions.MaxTimeSeries.get -> int
 Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsEndpointOptions.MaxTimeSeries.set -> void
 Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsEndpointOptions.MetricsEndpointOptions() -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.AspNetCoreHosting.get -> bool
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.AspNetCoreHosting.set -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.EgressIgnorePattern.get -> string?
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.EgressIgnorePattern.set -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.EventCounterEvents.get -> bool
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.EventCounterEvents.set -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.EventCounterIntervalSec.get -> int?
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.EventCounterIntervalSec.set -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.ExcludedMetrics.get -> System.Collections.Generic.IList<string!>!
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.GCEvents.get -> bool
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.GCEvents.set -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.HttpClientCore.get -> bool
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.HttpClientCore.set -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.HttpClientDesktop.get -> bool
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.HttpClientDesktop.set -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.IncludedMetrics.get -> System.Collections.Generic.IList<string!>!
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.IngressIgnorePattern.get -> string?
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.IngressIgnorePattern.set -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.MetricsObserverOptions() -> void
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.ThreadPoolEvents.get -> bool
+Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsObserverOptions.ThreadPoolEvents.set -> void
 Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsRequest
 Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsRequest.MetricName.get -> string!
 Steeltoe.Management.Endpoint.Actuators.Metrics.MetricsRequest.MetricsRequest(string! metricName, System.Collections.Generic.IList<System.Collections.Generic.KeyValuePair<string!, string!>>! tags) -> void
@@ -503,6 +540,10 @@ Steeltoe.Management.Endpoint.SpringBootAdminClient.SpringBootAdminClientOptions.
 Steeltoe.Management.Endpoint.SpringBootAdminClient.SpringBootAdminClientOptions.Url.set -> void
 Steeltoe.Management.Endpoint.SpringBootAdminClient.SpringBootAdminClientOptions.ValidateCertificates.get -> bool
 Steeltoe.Management.Endpoint.SpringBootAdminClient.SpringBootAdminClientOptions.ValidateCertificates.set -> void
+virtual Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.DiagnosticObserver.Dispose(bool disposing) -> void
+virtual Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.DiagnosticObserver.OnCompleted() -> void
+virtual Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.DiagnosticObserver.OnError(System.Exception! error) -> void
+virtual Steeltoe.Management.Endpoint.Actuators.Metrics.Diagnostics.DiagnosticObserver.OnNext(System.Collections.Generic.KeyValuePair<string!, object?> value) -> void
 virtual Steeltoe.Management.Endpoint.Configuration.ConfigureEndpointOptions<TOptions>.Configure(TOptions! options) -> void
 virtual Steeltoe.Management.Endpoint.Middleware.EndpointMiddleware<TArgument, TResult>.ShouldInvoke(Microsoft.AspNetCore.Http.PathString requestPath) -> bool
 virtual Steeltoe.Management.Endpoint.Middleware.EndpointMiddleware<TArgument, TResult>.WriteResponseAsync(TResult result, Microsoft.AspNetCore.Http.HttpContext! context, System.Threading.CancellationToken cancellationToken) -> System.Threading.Tasks.Task!
